fix: add error handling to jsonReviews (#1019)
* fix: add error handling to jsonFetch utility

This commit introduces error handling to the `jsonFetch` utility function. Now, when a fetch request fails, it will log the error and return null. Corresponding tests to validate this behavior have also been added.

* Revert "fix: add error handling to jsonFetch utility"

This reverts commit f64a26e3b080e8f6e5613cc248e8c3d581b2c125.

* feat: Improve error handling in fetchReviews function

This commit enhances error handling in the fetchReviews function in
fetch-data.ts. It now catches errors and logs them to the console. It also
returns a Promise even when reviewsJson[id] is available.