feat: add storybook jest for testing
This commit includes the addition of the `@storybook/jest` package to
the project's dependencies. This package is used for writing tests for
storybook stories. The `modal.stories.tsx` file was updated to include
a new test for the `ModalCite` component. The `yarn.lock` file was
also updated to include the newly added dependencies and their
versions.