feat: update msas field to use article subjects (#1047)
* feat: update msas field to use article subjects

The msas field in the reviewed preprints page is now populated with the article subjects. Additionally, a new field called subjects has been added to the enhanced-article type.

* feat: update EPP API image version in docker-compose

The docker image of the Enhanced Preprints API has been updated to a newer version.

* add failing test for the empty array

* feat: add conditional rendering to ArticleFlagList component

This commit introduces a conditional rendering in ArticleFlagList
component. If the msas prop length is zero, it now returns an empty
fragment.

---------

Co-authored-by: Will Byrne <will.alex.byrne@gmail.com>
Co-authored-by: Scott Aubrey <scott@aubrey.org.uk>