feat: make prepareAuthor function exportable and use it in [msid].page.ts (#1257)
* feat: make prepareAuthor function exportable and use it in [msid].page.ts

The `prepareAuthor` function from `reviewed-preprints.page.ts` is now
exportable. It is imported and utilized in `reviewed-preprints/[msid].page.ts`
to prepare the authors' metadata.

* Ensure that author names are coming through in indexContent

* Asjust the test to account for comma separated named

* feat: add test case for author with no family names in reviewed-preprints.page.test.ts

* feat: update authorLine in reviewed-preprints test data

This commit updates the authorLine field in the reviewed-preprints.page.test.ts
file to not include family names.

* feat: improve author name handling in reviewed-preprints.page.ts

Refactor the prepareAuthor function to handle cases where an author might not have a family name. This ensures no extra space is added when there is no family name.

---------

Co-authored-by: aishakapan <a.kapan@elifesciences.org>