feat: update reviewed preprints API to include first published date
This commit modifies the 'src/pages/api/reviewed-preprints.page.ts' and
'src/pages/api/reviewed-preprints/[msid].page.ts' files. It updates the
'enhancedArticleToReviewedPreprintItemResponse' function to include the
'firstPublished' date in the response. It also fetches the first published
version from the 'VersionSummary' when fetching a version in the
'reviewed-preprints/[msid].page.ts' API.