feat: update 'published' and 'reviewedDate' to use 'firstPublished' (#1509)
* feat: update 'published' and 'reviewedDate' to use 'firstPublished'

This commit updates the 'published' and 'reviewedDate' properties in
the 'reviewedPreprintSnippet' function to use 'firstPublished' instead
of 'published'. Also, it adds 'firstPublished' as a new property in
the 'EnhancedArticleNoContent' type. Additionally, it changes the
import path of 'Author' in the 'reviewed-preprints.page.test.ts' file.

* feat: update reviewed preprints API to include first published date

This commit modifies the 'src/pages/api/reviewed-preprints.page.ts' and
'src/pages/api/reviewed-preprints/[msid].page.ts' files. It updates the
'enhancedArticleToReviewedPreprintItemResponse' function to include the
'firstPublished' date in the response. It also fetches the first published
version from the 'VersionSummary' when fetching a version in the
'reviewed-preprints/[msid].page.ts' API.