feat: Set default automation flag and refactor browser tests (#1573)
* feat: Set default automation flag to true and remove local import and server section from README

This commit sets the default automation flag in the config to true. It also removes the section on running with a local import and server from the README file.

* Revert "feat: Set default automation flag to true and remove local import and server section from README"

This reverts commit fcbaec221779a15b1686eb0bc84d9bbe7c744335.

* feat: Set automationFlag to true in src/config.ts

* feat: Add new location to nginx config and test data

Added a new location '/api/preprints-no-content' to the nginx config and
created a new test data file 'preprints-no-content.json' with mock data for
testing.

* Make automation tests pass by refactoring mock api

* feat: update nginx config and content-header selector

Update .docker/nginx-browser-test.conf to remove redundant aliases. Modify
content-header.ts to adjust the DOI locator and change expect condition for
assertDOI function.

* feat: Improve readability of JSON data in test-utils

Reformatted JSON objects in preprints-no-content.json and 123.json for better readability. This includes better indentation and formatting of the authors and references sections. Also ensured newline at the end of 123.json.

---------

Co-authored-by: aishakapan <a.kapan@elifesciences.org>