feat: update citations API to use fetchVersion utility (#1586)
* feat: update citations API to use fetchVersion utility

Updated the '/api/citations/msid/bibtex' and '/api/citations/msid/ris' endpoints to use the 'fetchVersion' utility function instead of 'getManuscript' and 'getRppVersionDoi'. This change enhances the way we handle msid and error handling when a version is not retrieved. The doi is now directly accessed from the version object.

* feat: add citations download feature

This commit includes changes for downloading bibtex and ris citations.
Updates have been made to the nginx-browser-test.conf file to add new
location blocks for bibtex and ris citations. New test cases have been
added to the 'citations.spec.ts' file for testing the download of these
citations. The 'docker-compose.browsertest.yaml' file has been updated
to expose the API and client via proxy. New .bib and .ris files have
been added to the 'test-utils/data/preprints' directory for testing
purposes.

* feat: update test assertions in citations.spec.ts

This commit updates the test assertions in the browser-tests/citations.spec.ts file. The expected response bodies now include a period at the end of the sentence 'A study of world domination by genetically enhanced mice.'

* feat: add content-disposition header checks to browser tests

This commit introduces checks for the 'content-disposition' header in
the 'citations.spec.ts' browser tests. The header is expected to be
set to 'attachment; filename=123.bib' and 'attachment; filename=123.ris'
for bibtex and research-info-systems respectively.

* feat: change citation file extension from .bib to .ris

* feat: remove port exposure for API in browser test docker compose file

* feat: add tests for citation handlers and update deps

- Add 'node-mocks-http' to package.json
- Create tests for BibTeX and RIS citation handlers
- Remove redundant code from citation handlers
- Update citation data in test-utils
- Update yarn.lock with new dependencies

* feat: update citation identifiers in test files

In test files for bibtex and ris citations, updated the identifiers from '123' to '321'. This affects URLs, data returned, and fetch URLs in the tests.

* feat: add DOI check to citation API tests

Added tests in bibtex.page.test.ts and ris.page.test.ts to ensure the DOI
is contained within the response data.

* feat: update @types/node dependency in yarn.lock

The @types/node dependency in yarn.lock has been updated. The new version
is 20.11.3 and the old redundant entry for the same dependency has been
removed.

* Remove citations browser test