Switch backstop.js to use firefox as the visual regression engine
The output from chromium run in headless mode doesn't match what chromium shows when run headed.
I've been chasing down this bug on and off for weeks, and have just come to the conclusion that
chromium is not very good at taking reliable screenshots. Switching to firefox headless shows the
site as it is rendered when firefox is run in headed mode, so is easier to debug if it fails.
Which ultimately, is more valuable than a hard-to-debug Blink-based renderer failure.