feat: Update related content type translation in preprints page
This commit updates how the related content type is translated in the
reviewed preprints page. It now uses a more dynamic translation key
based on the content type, with a default value set for unspecified
content types. Additionally, a new translation key for 'Insight' type
has been added to the i18n configuration.