feat: refactor generateTimeline function in generate-timeline.ts
This commit refactors the generateTimeline function to streamline the code.
It renames 'isVersionPreprintSummary' to 'isPreprintVersionSummary' and
'isExternalPreprintSummary' to 'isExternalVersionSummary'. It also combines
the separate event push actions for different version types into a single one,
using a conditional statement to differentiate between 'PreprintVersionSummary'
and 'ExternalVersionSummary'. The 'link' url is also updated to include
a conditional check for version type.