feat: enhance previous version warning component style
The commit includes changes in both SCSS and TSX files of the previous
version warning component. The component is now wrapped inside a
'previous-version-warning-container' for better styling. Also, the
background color is now set in the container, and display property is
updated to 'flex'. The warning message is also now wrapped inside a
'div' instead of a 'span'.