Update styling for review process component [elifesciences/enhanced-preprints-issues#1114]
This commit includes minor visual changes to the review process component. Specifically, it adds margin and padding to the header and container respectively. The padding for the body and link elements is slightly increased. The body has also been given more bottom margin and a larger border radius for the pseudo-element before it. The width of this pseudo-element is slightly reduced and its horizontal position is shifted to the right.