Peer review process switch on (#1984)
* Add current version to ArticleReviewsTab and update test cases [elifesciences/enhanced-preprints-issues#1115]

This update includes passing the current version parameter to the ArticleReviewsTab component. It also involves changing the href URL in the review-timeline component, adding a new property to the i18n file, and updating various test cases to accommodate these changes.

* Update review timeline link and remove unused i18n entry [elifesciences/enhanced-preprints-issues#1115]

Updated the review timeline link to use the event's url directly instead of an i18n entry.
Removed the unused 'reviewed_preprints_peer_review_url' from i18n.ts.

* Remove unnecessary useTranslation hook in timeline.tsx [elifesciences/enhanced-preprints-issues#1115]

* Add margin and padding to review-process component [elifesciences/enhanced-preprints-issues#1115]

* Add href attribute to revision status in timeline [elifesciences/enhanced-preprints-issues#1113]

The 'Not revised' and 'Revised by authors' texts in the timeline now have
href attributes that link to the review process section. Changes are
reflected in the timeline test file and the timeline component file.

* Remove top padding from editors-and-reviewers component [elifesciences/enhanced-preprints-issues#1113]