Revert "Allow timeline css class names to be found in a code search by explicitly defining them rather than constructing them #2238"
The commit being reverted was a legitimate refactor: it changed the code without
changing the behaviour. However it introduced a merge conflict and so CI
wouldn't run, hence preventing the Chromatic check from running.
Reverting for now but the desired change is now understood so can be
reimplemented once the upstream changes have been merged in. Perhaps
when there's less traffic on the timeline.scss file.

This reverts commit ffa532737b4579ea5e26cd51e7d15846517d34a4.